Refactor logic and reduce component
